fix unique index of topic in pgsql database

+++ packages/watchmen-storage-postgresql/meta-scripts/18.0.0/00001-tenants_ddl.sql
@@ -1,1 +1,1 @@
-ALTER TABLE tenants ADD enable_ai DECIMAL(1) NOT NULL DEFAULT 0;
+ALTER TABLE tenants ADD enable_ai smallint NOT NULL DEFAULT 0;
